Feat : 자동결제 로직 추가

+++ backend/src/main/java/com/daou/project/model/CouponDto.java
@@ -5,8 +5,8 @@ import lombok.Setter;
 
 @Getter
 public class CouponDto {
-	private int couponNo;
-	private int userNo;
+	private long couponNo;
+	private long userNo;
 	private int type;
 	private char couponYn;
 }
