+++ backend/src/main/java/com/daou/project/model/PointDto.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PointDto {
-	private int pointNo;
-	private int userNo;
+	private long pointNo;
+	private long userNo;
 	private int savePoint;
 	private int balancePoint;
 	private String expirationDate;
