+++ backend/src/main/java/com/daou/project/model/ProductDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ProductDto {
-	private int productNo;
+	private long productNo;
 	private String productName;
 	private int productPrice;
 	private String productContent;
