+++ backend/src/main/java/com/daou/project/model/SaveMoneyDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class SaveMoneyDto {
-	private int savemoneyNo;
-	private int userNo;
+	private long savemoneyNo;
+	private long userNo;
 	private int saveMoney;
 }
